// User.View/src/app/models/user.ts
export interface User {
    id?: string | null;
    nome?: string | null;
    sobrenome?: string | null;
    email?: string | null;
    dataNascimento?: string | null;
    escolaridade?: string | null;
}

export class User {
    id?: string | null;
    nome?: string | null;
    sobrenome?: string | null;
    email?: string | null;
    dataNascimento?: string | null;
    escolaridade?: string | null;
}